Remove inefficient Regexp scan in compress_snappy_stream
Contrary to our expectation, we observed that logstash process consumes
more CPU time when using Snappy compression than when using GZ.

Profiling shows that the Regexp scans for splitting the input data into
chunks were using a large portion of the CPU time. We can change it to
a simple substring loop and remove the unnecessary overhead.